Move tests to microservice folder
