+++ .github/workflows/test.yml
@@ -24,6 +24,5 @@ jobs:
       - name: Build
         run: dotnet build --no-restore
 
-      - name: Test
-        run: dotnet test DonationService.Tests --no-build --verbosity normal
-
+      - name: Test DonationService
+        run: dotnet test Glense.Server/DonationService/Tests/DonationService.Tests.csproj --no-build --verbosity normal
