+++ Glense.sln
@@ -9,7 +9,7 @@ Project("{54A90642-561A-4BB1-A94E-469ADEE60C69}") = "glense.client", "glense.cli
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Glense.Server", "Glense.Server", "{127D1C99-BC8F-468B-80A6-96FDCFC1F790}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DonationService.Tests", "DonationService.Tests\DonationService.Tests.csproj", "{0FD4E6A0-9D52-4F8D-A67D-77A496568277}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DonationService.Tests", "Glense.Server\DonationService\Tests\DonationService.Tests.csproj", "{0FD4E6A0-9D52-4F8D-A67D-77A496568277}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DonationService", "Glense.Server\DonationService\DonationService.csproj", "{A1B2C3D4-E5F6-7890-ABCD-EF1234567890}"
 EndProject
